+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/_internal/configureDependencies.kt
@@ -13,10 +13,10 @@ import com.javiersc.hubdle.extensions.kotlin._internal.forKotlinSetsDependencies
 internal fun HubdleConfigurableExtension.configurableDependencies() {
     configurable {
         project.forKotlinSetsDependencies(
-            "test",
-            "commonTest",
-            "testFunctional",
-            "testIntegration",
+            TEST,
+            COMMON_TEST,
+            TEST_FUNCTIONAL,
+            TEST_INTEGRATION,
         ) {
             implementation(catalogDependency(ORG_JETBRAINS_KOTLIN_KOTLIN_TEST_MODULE))
             implementation(