+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/shared/HubdleKotlinSourceSetConfigurableExtension.kt
@@ -1,5 +1,10 @@
 package com.javiersc.hubdle.extensions.kotlin.shared
 
+import com.javiersc.hubdle.extensions._internal.MAIN
+import com.javiersc.hubdle.extensions._internal.TEST
+import com.javiersc.hubdle.extensions._internal.TEST_FIXTURES
+import com.javiersc.hubdle.extensions._internal.TEST_FUNCTIONAL
+import com.javiersc.hubdle.extensions._internal.TEST_INTEGRATION
 import com.javiersc.hubdle.extensions.apis.HubdleSourceSetConfigurableExtension
 import org.gradle.api.NamedDomainObjectProvider
 import org.gradle.api.Project
@@ -12,31 +17,31 @@ public abstract class HubdleKotlinSourceSetConfigurableExtension(
 
     override val main: NamedDomainObjectProvider<KotlinSourceSet>
         get() {
-            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName("main"))
-            return the<KotlinProjectExtension>().sourceSets.named(calculateName("main"))
+            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName(MAIN))
+            return the<KotlinProjectExtension>().sourceSets.named(calculateName(MAIN))
         }
 
     override val test: NamedDomainObjectProvider<KotlinSourceSet>
         get() {
-            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName("test"))
-            return the<KotlinProjectExtension>().sourceSets.named(calculateName("test"))
+            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName(TEST))
+            return the<KotlinProjectExtension>().sourceSets.named(calculateName(TEST))
         }
 
     override val testFunctional: NamedDomainObjectProvider<KotlinSourceSet>
         get() {
-            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName("testFunctional"))
-            return the<KotlinProjectExtension>().sourceSets.named(calculateName("testFunctional"))
+            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName(TEST_FUNCTIONAL))
+            return the<KotlinProjectExtension>().sourceSets.named(calculateName(TEST_FUNCTIONAL))
         }
 
     override val testIntegration: NamedDomainObjectProvider<KotlinSourceSet>
         get() {
-            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName("testIntegration"))
-            return the<KotlinProjectExtension>().sourceSets.named(calculateName("testIntegration"))
+            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName(TEST_INTEGRATION))
+            return the<KotlinProjectExtension>().sourceSets.named(calculateName(TEST_INTEGRATION))
         }
 
     override val testFixtures: NamedDomainObjectProvider<KotlinSourceSet>
         get() {
-            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName("testFixtures"))
-            return the<KotlinProjectExtension>().sourceSets.named(calculateName("testFixtures"))
+            the<KotlinProjectExtension>().sourceSets.maybeCreate(calculateName(TEST_FIXTURES))
+            return the<KotlinProjectExtension>().sourceSets.named(calculateName(TEST_FIXTURES))
         }
 }